+++ src/main.rs
@@ -16,14 +16,12 @@
 
 #![feature(duration_checked_float)]
 
-use sw::{FatalError, Logger, UserError};
+use sw::color::red;
+use sw::error::FatalError;
+use sw::state::{Command, State};
 
-use libsw::Stopwatch;
-use log::{debug, error, info, trace, warn};
-
-use std::io::{self, BufRead, BufWriter, Read, Write};
+use std::io::{self, BufWriter, Write};
 use std::process::ExitCode;
-use std::time::Duration;
 
 fn main() -> ExitCode {
     if let Err(error) = try_main() {
@@ -35,7 +33,6 @@ fn main() -> ExitCode {
 }
 
 fn try_main() -> Result<(), FatalError> {
-    Logger::init().unwrap();
     print_splash()?;
 
     let mut state = State::new();
@@ -45,11 +42,11 @@ fn try_main() -> Result<(), FatalError> {
         // respond to command
         match Command::read(&state.name, state.sw.is_running())? {
             Ok(command) => {
-                if state.update(command, &mut stdout)? {
+                if state.update(&command, &mut stdout)? {
                     return Ok(());
                 }
             }
-            Err(error) => error!("{}", error),
+            Err(error) => red(error)?,
         }
 
         writeln!(stdout)?;
@@ -64,24 +61,6 @@ fn try_main() -> Result<(), FatalError> {
             state.since_stop.start().unwrap();
         }
     }
-}
-
-fn readln(msg: &str) -> Result<String, FatalError> {
-    const READ_LIMIT: u64 = 128;
-
-    let mut stdout = io::stdout();
-    let stdin = io::stdin().lock();
-    let mut input = String::new();
-
-    // print prompt (must flush)
-    write!(stdout, "{}", msg)?;
-    stdout.flush()?;
-
-    // read a limited number of bytes from stdin
-    stdin.take(READ_LIMIT).read_line(&mut input)?;
-
-    // trim whitespace and escape weird characters
-    Ok(input.trim().escape_default().to_string())
 }
 
 fn print_splash() -> Result<(), FatalError> {
@@ -100,225 +79,4 @@ fn print_splash() -> Result<(), FatalError> {
     stderr.flush()?;
 
     Ok(())
-}
-
-struct State {
-    sw: Stopwatch,
-    since_stop: Stopwatch,
-    name: String,
-    prec: usize,
-}
-
-impl State {
-    const PRECISION_INIT: usize = 2;
-    const PRECISION_MAX: usize = 9; // nanosecond precision
-
-    pub fn new() -> Self {
-        Self {
-            sw: Stopwatch::new(),
-            since_stop: Stopwatch::new_started(),
-            name: String::new(),
-            prec: Self::PRECISION_INIT,
-        }
-    }
-
-    /// Sets the precision of displayed values. If the precision was clamped, it
-    /// returns what it was clamped to.
-    pub fn set_precision(&mut self, prec: usize) -> Option<usize> {
-        if prec <= Self::PRECISION_MAX {
-            self.prec = prec;
-            None
-        } else {
-            self.prec = Self::PRECISION_MAX;
-            Some(Self::PRECISION_MAX)
-        }
-    }
-
-    /// Updates the state with a [`Command`], returning [true] if the command
-    /// was [`Quit`](Command::Quit).
-    pub fn update<W: Write>(
-        &mut self,
-        command: Command,
-        stdout: &mut W,
-    ) -> Result<bool, FatalError> {
-        match command {
-            Command::Help => {
-                writeln!(stdout, "| command | description           |")?;
-                writeln!(stdout, "| ------- | --------------------- |")?;
-                writeln!(stdout, "| h       | print this message    |")?;
-                writeln!(stdout, "| <enter> | display elapsed time  |")?;
-                writeln!(stdout, "| s       | toggle stopwatch      |")?;
-                writeln!(stdout, "| r       | reset stopwatch       |")?;
-                writeln!(stdout, "| c       | change elapsed time   |")?;
-                writeln!(stdout, "| o       | offset elapsed time   |")?;
-                writeln!(stdout, "| n       | name stopwatch        |")?;
-                writeln!(stdout, "| p       | set display precision |")?;
-                writeln!(stdout, "| l       | print license info    |")?;
-                writeln!(stdout, "| q       | Abandon all Data      |")?;
-            }
-
-            Command::Display => {
-                let elapsed = self.sw.elapsed().as_secs_f32();
-
-                // display time elapsed in different units
-                writeln!(stdout, "{:.*} seconds", self.prec, elapsed)?;
-                writeln!(stdout, "{:.*} minutes", self.prec, elapsed / 60.0)?;
-                writeln!(stdout, "{:.*} hours", self.prec, elapsed / 60.0 / 60.0)?;
-
-                stdout.flush()?;
-
-                // indicate status
-                if self.sw.is_running() {
-                    debug!("running");
-                } else {
-                    warn!("stopped");
-                }
-            }
-
-            Command::Toggle => {
-                self.sw.toggle();
-                if self.sw.is_running() {
-                    info!("started stopwatch");
-                    trace!(
-                        "{:.*} seconds since stopped",
-                        self.prec,
-                        self.since_stop.elapsed().as_secs_f32()
-                    );
-                } else {
-                    info!("stopped stopwatch");
-                }
-            }
-
-            Command::Reset => {
-                self.sw.reset();
-                info!("reset stopwatch");
-            }
-
-            Command::Change => match read_duration("new value? ")? {
-                Ok((dur, is_neg)) => {
-                    if is_neg {
-                        error!("{}", UserError::NegativeDuration);
-                    } else {
-                        self.sw.set(dur);
-                        info!("updated elapsed time");
-                    }
-                }
-                Err(error) => error!("{}", error),
-            },
-
-            Command::Offset => match read_duration("offset by? ")? {
-                Ok((dur, is_neg)) => {
-                    if is_neg {
-                        self.sw -= dur;
-                        info!("subtracted from elapsed time");
-                    } else {
-                        self.sw += dur;
-                        info!("added to elapsed time");
-                    }
-                }
-                Err(error) => error!("{}", error),
-            },
-
-            Command::Name => {
-                self.name = readln("new name? ")?;
-                if self.name.is_empty() {
-                    info!("cleared stopwatch name");
-                } else {
-                    info!("updated stopwatch name");
-                }
-            }
-
-            Command::Precision => match readln("new precision? ")?.parse::<usize>() {
-                Ok(int) => {
-                    if let Some(clamped) = self.set_precision(int) {
-                        warn!("precision clamped to {}", clamped);
-                    } else {
-                        info!("updated precision");
-                    }
-                }
-                Err(error) => error!("{}", UserError::InvalidInt(error)),
-            },
-
-            Command::License => {
-                writeln!(stdout, "copyright (C) 2022  {}", env!("CARGO_PKG_AUTHORS"))?;
-                writeln!(stdout, "licensed under {}", env!("CARGO_PKG_LICENSE"))?;
-            }
-
-            Command::Quit => return Ok(true),
-        };
-
-        Ok(false)
-    }
-}
-
-enum Command {
-    Help,
-    Display,
-    Toggle,
-    Reset,
-    Change,
-    Offset,
-    Name,
-    Precision,
-    License,
-    Quit,
-}
-
-impl Command {
-    pub fn read(msg: &str, running: bool) -> Result<Result<Self, UserError>, FatalError> {
-        let prompt = format!("{} {} ", msg, if running { '>' } else { '<' });
-        match readln(&prompt)?.to_lowercase().as_ref() {
-            "h" => Ok(Ok(Self::Help)),
-            "" => Ok(Ok(Self::Display)),
-            "s" => Ok(Ok(Self::Toggle)),
-            "r" => Ok(Ok(Self::Reset)),
-            "c" => Ok(Ok(Self::Change)),
-            "o" => Ok(Ok(Self::Offset)),
-            "n" => Ok(Ok(Self::Name)),
-            "p" => Ok(Ok(Self::Precision)),
-            "l" => Ok(Ok(Self::License)),
-            "q" => Ok(Ok(Self::Quit)),
-            other => Ok(Err(UserError::UnrecognizedCommand(other.into()))),
-        }
-    }
-}
-
-enum Unit {
-    Seconds,
-    Minutes,
-    Hours,
-}
-
-impl Unit {
-    pub fn from_str(s: &str) -> Result<Self, UserError> {
-        match s {
-            "s" => Ok(Self::Seconds),
-            "m" => Ok(Self::Minutes),
-            "h" => Ok(Self::Hours),
-            s => Err(UserError::UnrecognizedUnit(s.into())),
-        }
-    }
-}
-
-fn read_duration(msg: &str) -> Result<Result<(Duration, bool), UserError>, FatalError> {
-    let mut s = readln(msg)?;
-    let try_unit = s.pop().map(|chr| chr.to_string()).unwrap_or_default();
-    match s.parse::<f64>() {
-        Ok(scalar) => {
-            let unit = match Unit::from_str(&try_unit) {
-                Ok(unit) => unit,
-                Err(error) => return Ok(Err(error)),
-            };
-            let secs = match unit {
-                Unit::Seconds => scalar,
-                Unit::Minutes => scalar * 60.0,
-                Unit::Hours => scalar * 60.0 * 60.0,
-            };
-            match Duration::try_from_secs_f64(secs.abs()) {
-                Ok(duration) => Ok(Ok((duration, secs.is_sign_negative()))),
-                Err(error) => Ok(Err(UserError::InvalidDuration(error))),
-            }
-        }
-        Err(error) => Ok(Err(UserError::InvalidFloat(error))),
-    }
 }