+++ src/state.rs
@@ -1,0 +1,284 @@
+// sw: terminal stopwatch
+// Copyright (C) 2022  Ula Shipman
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
+//! `sw` state.
+
+use crate::color::{green, grey, magenta, red, yellow};
+use crate::error::{FatalError, UserError};
+
+use libsw::Stopwatch;
+
+use core::time::Duration;
+use std::io::{self, BufRead, Read, Write};
+
+/// Stopwatch state.
+pub struct State {
+    /// Main [`Stopwatch`].
+    pub sw: Stopwatch,
+    /// Records the time elapsed while `sw` is stopped.
+    pub since_stop: Stopwatch,
+    /// Name of the stopwatch.
+    pub name: String,
+    prec: usize,
+}
+
+impl State {
+    const PRECISION_INIT: usize = 2;
+    const PRECISION_MAX: usize = 9; // nanosecond precision
+
+    /// Returns a new [`State`].
+    #[allow(clippy::new_without_default)]
+    #[must_use]
+    pub fn new() -> Self {
+        Self {
+            sw: Stopwatch::new(),
+            since_stop: Stopwatch::new_started(),
+            name: String::new(),
+            prec: Self::PRECISION_INIT,
+        }
+    }
+
+    /// Sets the precision of displayed values. If the precision was clamped, it
+    /// returns what it was clamped to.
+    pub fn set_precision(&mut self, prec: usize) -> Option<usize> {
+        if prec <= Self::PRECISION_MAX {
+            self.prec = prec;
+            None
+        } else {
+            self.prec = Self::PRECISION_MAX;
+            Some(Self::PRECISION_MAX)
+        }
+    }
+
+    /// Updates the state with a [`Command`], returning [true] if the command
+    /// was [`Quit`](Command::Quit).
+    ///
+    /// # Errors
+    ///
+    /// - The command may be invalid
+    /// - Reading and writing to the terminal may fail
+    pub fn update<W: Write>(
+        &mut self,
+        command: &Command,
+        stdout: &mut W,
+    ) -> Result<bool, FatalError> {
+        match command {
+            Command::Help => {
+                writeln!(stdout, "| command | description           |")?;
+                writeln!(stdout, "| ------- | --------------------- |")?;
+                writeln!(stdout, "| h       | print this message    |")?;
+                writeln!(stdout, "| <enter> | display elapsed time  |")?;
+                writeln!(stdout, "| s       | toggle stopwatch      |")?;
+                writeln!(stdout, "| r       | reset stopwatch       |")?;
+                writeln!(stdout, "| c       | change elapsed time   |")?;
+                writeln!(stdout, "| o       | offset elapsed time   |")?;
+                writeln!(stdout, "| n       | name stopwatch        |")?;
+                writeln!(stdout, "| p       | set display precision |")?;
+                writeln!(stdout, "| l       | print license info    |")?;
+                writeln!(stdout, "| q       | Abandon all Data      |")?;
+            }
+
+            Command::Display => {
+                let elapsed = self.sw.elapsed().as_secs_f32();
+
+                // display time elapsed in different units
+                writeln!(stdout, "{:.*} seconds", self.prec, elapsed)?;
+                writeln!(stdout, "{:.*} minutes", self.prec, elapsed / 60.0)?;
+                writeln!(stdout, "{:.*} hours", self.prec, elapsed / 60.0 / 60.0)?;
+
+                stdout.flush()?;
+
+                // indicate status
+                if self.sw.is_running() {
+                    green("running")?;
+                } else {
+                    yellow("stopped")?;
+                }
+            }
+
+            Command::Toggle => {
+                self.sw.toggle();
+                if self.sw.is_running() {
+                    magenta("started stopwatch")?;
+                    grey(format!(
+                        "{:.*} seconds since stopped",
+                        self.prec,
+                        self.since_stop.elapsed().as_secs_f32(),
+                    ))?;
+                } else {
+                    magenta("stopped stopwatch")?;
+                }
+            }
+
+            Command::Reset => {
+                self.sw.reset();
+                magenta("reset stopwatch")?;
+            }
+
+            Command::Change => match read_duration("new value? ")? {
+                Ok((dur, is_neg)) => {
+                    if is_neg {
+                        red(UserError::NegativeDuration)?;
+                    } else {
+                        self.sw.set(dur);
+                        magenta("updated elapsed time")?;
+                    }
+                }
+                Err(error) => red(error)?,
+            },
+
+            Command::Offset => match read_duration("offset by? ")? {
+                Ok((dur, is_neg)) => {
+                    if is_neg {
+                        self.sw -= dur;
+                        magenta("subtracted from elapsed time")?;
+                    } else {
+                        self.sw += dur;
+                        magenta("added to elapsed time")?;
+                    }
+                }
+                Err(error) => red(error)?,
+            },
+
+            Command::Name => {
+                self.name = readln("new name? ")?;
+                if self.name.is_empty() {
+                    magenta("cleared stopwatch name")?;
+                } else {
+                    magenta("updated stopwatch name")?;
+                }
+            }
+
+            Command::Precision => match readln("new precision? ")?.parse::<usize>() {
+                Ok(int) => {
+                    if let Some(clamped) = self.set_precision(int) {
+                        yellow(format!("precision clamped to {}", clamped))?;
+                    } else {
+                        magenta("updated precision")?;
+                    }
+                }
+                Err(error) => red(UserError::InvalidInt(error))?,
+            },
+
+            Command::License => {
+                writeln!(stdout, "copyright (C) 2022  {}", env!("CARGO_PKG_AUTHORS"))?;
+                writeln!(stdout, "licensed under {}", env!("CARGO_PKG_LICENSE"))?;
+            }
+
+            Command::Quit => return Ok(true),
+        }
+
+        Ok(false)
+    }
+}
+
+/// Enumeration over commands which can be executed by a [`State`].
+#[allow(missing_docs)]
+pub enum Command {
+    Help,
+    Display,
+    Toggle,
+    Reset,
+    Change,
+    Offset,
+    Name,
+    Precision,
+    License,
+    Quit,
+}
+
+impl Command {
+    /// Read a `Command` from the user.
+    ///
+    /// # Errors
+    ///
+    /// - Reading from `stdin` may fail
+    /// - The user may enter an invalid command
+    pub fn read(msg: &str, running: bool) -> Result<Result<Self, UserError>, FatalError> {
+        let prompt = format!("{} {} ", msg, if running { '>' } else { '<' });
+        match readln(&prompt)?.to_lowercase().as_ref() {
+            "h" => Ok(Ok(Self::Help)),
+            "" => Ok(Ok(Self::Display)),
+            "s" => Ok(Ok(Self::Toggle)),
+            "r" => Ok(Ok(Self::Reset)),
+            "c" => Ok(Ok(Self::Change)),
+            "o" => Ok(Ok(Self::Offset)),
+            "n" => Ok(Ok(Self::Name)),
+            "p" => Ok(Ok(Self::Precision)),
+            "l" => Ok(Ok(Self::License)),
+            "q" => Ok(Ok(Self::Quit)),
+            other => Ok(Err(UserError::UnrecognizedCommand(other.into()))),
+        }
+    }
+}
+
+fn read_duration(msg: &str) -> Result<Result<(Duration, bool), UserError>, FatalError> {
+    enum Unit {
+        Seconds,
+        Minutes,
+        Hours,
+    }
+
+    impl Unit {
+        fn from_str(s: &str) -> Result<Self, UserError> {
+            match s {
+                "s" => Ok(Self::Seconds),
+                "m" => Ok(Self::Minutes),
+                "h" => Ok(Self::Hours),
+                s => Err(UserError::UnrecognizedUnit(s.into())),
+            }
+        }
+    }
+
+    let mut s = readln(msg)?;
+    let try_unit = s.pop().map(|chr| chr.to_string()).unwrap_or_default();
+    match s.parse::<f64>() {
+        Ok(scalar) => {
+            let unit = match Unit::from_str(&try_unit) {
+                Ok(unit) => unit,
+                Err(error) => return Ok(Err(error)),
+            };
+            let secs = match unit {
+                Unit::Seconds => scalar,
+                Unit::Minutes => scalar * 60.0,
+                Unit::Hours => scalar * 60.0 * 60.0,
+            };
+            match Duration::try_from_secs_f64(secs.abs()) {
+                Ok(duration) => Ok(Ok((duration, secs.is_sign_negative()))),
+                Err(error) => Ok(Err(UserError::InvalidDuration(error))),
+            }
+        }
+        Err(error) => Ok(Err(UserError::InvalidFloat(error))),
+    }
+}
+
+fn readln(msg: &str) -> Result<String, FatalError> {
+    const READ_LIMIT: u64 = 128;
+
+    let mut stdout = io::stdout();
+    let stdin = io::stdin().lock();
+    let mut input = String::new();
+
+    // print prompt (must flush)
+    write!(stdout, "{}", msg)?;
+    stdout.flush()?;
+
+    // read a limited number of bytes from stdin
+    stdin.take(READ_LIMIT).read_line(&mut input)?;
+
+    // trim whitespace and escape weird characters
+    Ok(input.trim().escape_default().to_string())
+}